Extract Metadata from RichEvent
- Extract Metadata from RichEvent
- Move large image and priority booking settings *out* of Metadata
- Move social hashtag *in* to Metadata
